Refactor: PalladioResourceRepository location
