+++ palladio-builder/src/main/java/de/fabiankeller/palladio/builder/PcmBuilder.java
@@ -10,9 +10,9 @@ import de.fabiankeller.palladio.builder.system.SystemBuilder;
 import de.fabiankeller.palladio.builder.system.impl.SystemBuilderImpl;
 import de.fabiankeller.palladio.builder.usage.UsageBuilder;
 import de.fabiankeller.palladio.builder.usage.impl.UsageBuilderImpl;
-import de.fabiankeller.palladio.builder.util.PalladioResourceRepository;
-import de.fabiankeller.palladio.builder.util.impl.PalladioResourceRepositoryImpl;
 import de.fabiankeller.palladio.environment.PCMResourceSetPartitionFactory;
+import de.fabiankeller.palladio.environment.util.PalladioResourceRepository;
+import de.fabiankeller.palladio.environment.util.PalladioResourceRepositoryImpl;
 import org.eclipse.emf.common.util.URI;
 import org.palladiosimulator.analyzer.workflow.blackboard.PCMResourceSetPartition;
 import org.palladiosimulator.solver.models.PCMInstance;
