+++ palladio-builder/src/main/java/de/fabiankeller/palladio/builder/repository/seff/impl/InternalActionBuilderImpl.java
@@ -3,8 +3,8 @@ package de.fabiankeller.palladio.builder.repository.seff.impl;
 import de.fabiankeller.palladio.builder.repository.failure.SoftwareInducedFailureBuilder;
 import de.fabiankeller.palladio.builder.repository.seff.InternalActionBuilder;
 import de.fabiankeller.palladio.builder.repository.seff.ResourceDemandBuilder;
-import de.fabiankeller.palladio.builder.util.PalladioResourceRepository;
 import de.fabiankeller.palladio.builder.util.RandomVariableFactory;
+import de.fabiankeller.palladio.environment.util.PalladioResourceRepository;
 import org.palladiosimulator.pcm.reliability.InternalFailureOccurrenceDescription;
 import org.palladiosimulator.pcm.reliability.ReliabilityFactory;
 import org.palladiosimulator.pcm.resourcetype.ProcessingResourceType;
