+++ palladio-builder/src/main/java/de/fabiankeller/palladio/builder/resourceenvironment/impl/LinkBuilderImpl.java
@@ -4,8 +4,8 @@ import de.fabiankeller.palladio.builder.AbstractHierarchicalBuilder;
 import de.fabiankeller.palladio.builder.resourceenvironment.ContainerBuilder;
 import de.fabiankeller.palladio.builder.resourceenvironment.LinkBuilder;
 import de.fabiankeller.palladio.builder.resourceenvironment.ResourceEnvironmentBuilder;
-import de.fabiankeller.palladio.builder.util.PalladioResourceRepository;
 import de.fabiankeller.palladio.builder.util.RandomVariableFactory;
+import de.fabiankeller.palladio.environment.util.PalladioResourceRepository;
 import org.palladiosimulator.pcm.core.PCMRandomVariable;
 import org.palladiosimulator.pcm.resourceenvironment.CommunicationLinkResourceSpecification;
 import org.palladiosimulator.pcm.resourceenvironment.LinkingResource;
