+++ palladio-builder/src/test/java/de/fabiankeller/palladio/builder/repository/seff/impl/ResourceDemandBuilderImplTest.java
@@ -4,9 +4,9 @@ import de.fabiankeller.palladio.builder.BuilderException;
 import de.fabiankeller.palladio.builder.repository.*;
 import de.fabiankeller.palladio.builder.repository.impl.RepositoryBuilderImpl;
 import de.fabiankeller.palladio.builder.repository.seff.ResourceDemandBuilder;
-import de.fabiankeller.palladio.builder.util.PalladioResourceRepository;
 import de.fabiankeller.palladio.config.EnvironmentConfig;
 import de.fabiankeller.palladio.environment.PalladioEclipseEnvironment;
+import de.fabiankeller.palladio.environment.util.PalladioResourceRepository;
 import org.eclipse.emf.common.util.EList;
 import org.junit.Before;
 import org.junit.Test;
